chore: refactor messages and stories
