+++ src/stories/buttons/PrimaryButton/PrimaryButton.stories.tsx
@@ -1,18 +1,18 @@
 import type { Meta, StoryObj } from "@storybook/react";
-
 import { PrimaryButton } from "./PrimaryButton";
 
-const meta = {
+const meta: Meta<typeof PrimaryButton> = {
   title: "Buttons/PrimaryButton",
   component: PrimaryButton,
   parameters: {
     layout: "centered",
   },
   tags: ["autodocs"],
-} as Meta<typeof PrimaryButton>;
+};
 
 export default meta;
-type Story = StoryObj<typeof meta>;
+
+type Story = StoryObj<typeof PrimaryButton>;
 
 export const Fill: Story = {
   args: {
