+++ src/stories/buttons/PrimaryButton/PrimaryButton.tsx
@@ -19,7 +19,6 @@ export const PrimaryButton = ({
   mode = "fill",
   isLocked,
   label,
-  ...props
 }: PrimaryButtonProps) => {
   const toggleColor = primary
     ? styles.fill
@@ -30,7 +29,6 @@ export const PrimaryButton = ({
     <button
       type={buttonType}
       className={[styles.primaryButton, styles[size], toggleColor].join(" ")}
-      {...props}
     >
       {isLocked ? "Locked" : isLocked === undefined ? label : "Unlocked"}
     </button>
