+++ src/stories/buttons/SubmitButton/SubmitButton.stories.tsx
@@ -2,17 +2,17 @@ import type { Meta, StoryObj } from "@storybook/react";
 
 import { SubmitButton } from "./SubmitButton";
 
-const meta = {
+const meta: Meta<typeof SubmitButton> = {
   title: "Buttons/SubmitButton",
   component: SubmitButton,
   parameters: {
     layout: "centered",
   },
   tags: ["autodocs"],
-} as Meta<typeof SubmitButton>;
+};
 
 export default meta;
-type Story = StoryObj<typeof meta>;
+type Story = StoryObj<typeof SubmitButton>;
 
 export const Submit: Story = {
   args: {
