+++ src/stories/forms/QuestionForm/QuestionForm.stories.tsx
@@ -1,18 +1,17 @@
-import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { QuestionForm } from "./QuestionForm";
 
-const meta = {
+const meta: Meta<typeof QuestionForm> = {
   title: "Forms/QuestionForm",
   component: QuestionForm,
   parameters: {
     layout: "centered",
   },
   tags: ["autodocs"],
-} as Meta<typeof QuestionForm>;
+};
 
 export default meta;
-type Story = StoryObj<typeof meta>;
+type Story = StoryObj<typeof QuestionForm>;
 
 export const Default: Story = {};
