+++ src/stories/forms/QuizForm/QuizForm.stories.tsx
@@ -1,18 +1,17 @@
-import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { Quiz } from "./QuizForm";
 
-const meta = {
+const meta: Meta<typeof Quiz> = {
   title: "Forms/QuizForm",
   component: Quiz,
   parameters: {
     layout: "centered",
   },
   tags: ["autodocs"],
-} as Meta<typeof Quiz>;
+};
 
 export default meta;
-type Story = StoryObj<typeof meta>;
+type Story = StoryObj<typeof Quiz>;
 
 export const Default: Story = {};
