+++ src/stories/index.ts
@@ -1,3 +1,5 @@
 export * from "./buttons";
 export * from "./header";
 export * from "./page";
+export * from "./forms";
+export * from "./tool-tips";
