+++ src/stories/tool-tips/QuestionIconToolTip/QuestionIconToolTip.stories.tsx
@@ -2,16 +2,16 @@ import type { Meta, StoryObj } from "@storybook/react";
 
 import { QuestionIconToolTip } from "./QuestionIconToolTip";
 
-const meta = {
+const meta: Meta<typeof QuestionIconToolTip> = {
   title: "ToolTips/QuestionIconToolTip",
   component: QuestionIconToolTip,
   parameters: {
     layout: "centered",
   },
   tags: ["autodocs"],
-} as Meta<typeof QuestionIconToolTip>;
+};
 
 export default meta;
-type Story = StoryObj<typeof meta>;
+type Story = StoryObj<typeof QuestionIconToolTip>;
 
 export const Default: Story = {};
